refactor: group by controller category
- 기존 CROUD에서 테이블 정보를 불러오는 컨트롤러가 추가되었기 때문에,
데이터를 쓰고, 읽고, 삭제, 수정하는 컨트롤러와 단순한 정보를 불러오
는 컨트롤러를 분리 (팀원의견)
- elasitc search는 비동기 이벤트가 있기 때문에, 정상적으로 테스트가
되지 않고, 추후 수정 예정
- close #29
