+++ .github/workflows/server-api-deploy.yml
@@ -25,7 +25,7 @@ jobs:
       run: |
         echo "##[set-output name=list;]$(echo $(git remote -v) | cut -d' ' -f2) ${{ steps.BaseRef.outputs.branch }} $GITHUB_REPOSITORY"
       id: Arguments
-    - name: Decrypt Environment and Send Data
+    - name: Decrypt Environment and build
       env:
         LARGE_SECRET_PASSPHRASE: ${{ secrets.PRIVATEKEY }}
         host: ${{ secrets[steps.HostName.outputs.host] }}
@@ -44,15 +44,9 @@ jobs:
         else
           touch ../../.env
         fi
-    - name: copy file via ssh password
-      uses: appleboy/scp-action@master
-      with:
-        host: ${{ secrets[steps.HostName.outputs.host] }}
-        username: ${{ secrets.USERNAME }}
-        password: ${{ secrets.PASSWORD }}
-        port: ${{ secrets.PORT }}
-        source: ".env"
-        target: "~/"
+        yarn build
+        cd ../../
+        sshpass -p$passowrd scp -P $port -o StrictHostKeyChecking=no . $username@$password:~/
     - name: deploy to server
       uses: appleboy/ssh-action@master
       env:
