+++ apis/product/package.json
@@ -9,7 +9,7 @@
     ]
   },
   "scripts": {
-    "test": "jest --runInBand",
+    "test": "jest --runInBand --setupFiles dotenv/config --forceExit",
     "precommit": "lint-staged && yarn test",
     "dev:start": "babel-node bin/www.js",
     "build": "babel ./ -d build --copy-files '*.json'",
