+++ apis/product/routes/controller/info.js
@@ -1,0 +1,8 @@
+import { getProductSchemaByKey } from '../../core';
+
+const getProductSchemaController = (req, res) => {
+  const list = getProductSchemaByKey('category').enumValues;
+  res.json(list);
+};
+
+export default getProductSchemaController;
